+++ client/common/App.js
@@ -29,6 +29,13 @@ export class App extends Component {
     }
   }
 
+  // React 16 新增方法，用来处理错误边界，可以捕获整个子组件树内发生的任何异常
+  componentDidCatch(error, errorInfo) {
+    // 可以打印异常，或者往后端日志中发送异常，方便定位跟踪
+    console.info(errorInfo);
+    console.error(errorInfo);
+  }
+
   // toast 组件
   renderToast() {
     const {