+++ client/utils/config.js
@@ -1,4 +1,4 @@
 // url 二级路由，取 server 端的值
-const {urlContext} = require('../../server/config');
+const {urlContext} = require('../../server/config/base');
 
 module.exports = {urlContext};