+++ client/utils/main-css.js
@@ -1,6 +1,0 @@
-import classNames from 'classnames/bind';
-import styles from '../common/scss/main.scss';
-
-const mainCss = classNames.bind(styles);
-
-export default mainCss;